+++ models/networks/basic/nn3.py
@@ -1,9 +1,6 @@
 import torch.nn as nn
 import torch.nn.functional as F
 
-import torch.nn as nn
-import torch.nn.functional as F
-
 
 class ComplexNN(nn.Module):
     """