+++ models/networks/transformers/__init__.py
@@ -1,0 +1,3 @@
+from .lorentz_transformer import (
+    LorentzInvariantAttention as LorentzAtt
+)